+++ src/routes/productsRoutes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const productsRouter = express.Router();
 const productsController = require('../controllers/productsController');
-const upload = require('../middlewares/multer');
+const upload = require('../middlewares/products-multer');
 
 /* PRODUCT LIST */
 productsRouter.get('/', productsController.index);
